+++ facade/src/main/scala/react/semanticui/modules/tab/Tab.scala
@@ -9,7 +9,8 @@ import japgolly.scalajs.react.component.Js.RawMounted
 import japgolly.scalajs.react.component.Js.UnmountedMapped
 import japgolly.scalajs.react.internal.Effect.Id
 import react.common.syntax._
-// import react.semanticui.elements.menu.Menu.MenuPros
+import react.semanticui.collections.menu.Menu
+import react.semanticui.collections.grid.Grid
 import react.semanticui._
 
 object Tab {
@@ -39,14 +40,13 @@ object Tab {
     var activeIndex: js.UndefOr[JsNumber | String]
 
     /** Shorthand props for the Menu. */
-    // var menu: js.UndefOr[MenuProps]
+    var menu: js.UndefOr[Menu.MenuProps]
 
     /** Align vertical menu */
-    // var menuPosition: js.UndefOr['left' | 'right']
     var menuPosition: js.UndefOr[String]
 
     /** Shorthand props for the Grid. */
-    // var grid: js.UndefOr[any]
+    var grid: js.UndefOr[Grid.GridProps]
 
     /**
       * Called on tab change.
@@ -82,23 +82,23 @@ object Tab {
   }
 
   def props(
-    as:                 js.UndefOr[AsC]               = js.undefined,
-    defaultActiveIndex: js.UndefOr[JsNumber | String] = js.undefined,
-    activeIndex:        js.UndefOr[JsNumber | String] = js.undefined,
-    // menu:        js.UndefOr[MenuProps]                              = js.undefined,
-    menuPosition:     js.UndefOr[TabMenuPosition]      = js.undefined,
-    onTabChangeE:     js.UndefOr[OnTabChange]          = js.undefined,
-    onTabChange:      js.UndefOr[TabProps => Callback] = js.undefined,
-    renderActiveOnly: js.UndefOr[Boolean]              = js.undefined,
+    as:                 js.UndefOr[AsC]                  = js.undefined,
+    defaultActiveIndex: js.UndefOr[JsNumber | String]    = js.undefined,
+    activeIndex:        js.UndefOr[JsNumber | String]    = js.undefined,
+    menu:               js.UndefOr[Menu.MenuProps]       = js.undefined,
+    menuPosition:       js.UndefOr[TabMenuPosition]      = js.undefined,
+    grid:               js.UndefOr[Grid.GridProps]       = js.undefined,
+    onTabChangeE:       js.UndefOr[OnTabChange]          = js.undefined,
+    onTabChange:        js.UndefOr[TabProps => Callback] = js.undefined,
+    renderActiveOnly:   js.UndefOr[Boolean]              = js.undefined
   ): TabProps = {
     val p = as.toJsObject[TabProps]
     p.as                 = as.toJs
     p.defaultActiveIndex = defaultActiveIndex
     p.activeIndex        = activeIndex
-    // p.menu= menu
-    p.menuPosition = menuPosition.toJs
-    // p.onBlur               = (onBlurE, onBlur).toJs
-    // p.onChange             = onChange.toJs
+    p.menu               = menu
+    p.menuPosition       = menuPosition.toJs
+    p.grid               = grid
     p.onTabChange =
       onTabChangeE.toJs.orElse(onTabChange.map(t => (_: ReactEvent, b: TabProps) => t(b).runNow))
     p.renderActiveOnly = renderActiveOnly