+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changes
 
+## 0.1.5
+
+* Upgrade to react-semantic-ui 0.87.2
+* Support Item
+* Add missing Tab properites menu and grid
+* Support ModalHeader, ModalDescription, ModalContent, ModalActions and Modal
+
 ## 0.1.4
 
 * Update scalajs-react-common